allow larger canvases
the multiply factor was too large for 1000*1000, so changed mutliplier.

Notes:
decided to go with a single box that displays current color picked.
have a rectangle under the canvas for deleting, going left and right to next and previous frame, showing frame number, etc..
fix weird canvas scaling with larger canvas sizes. maybe ratio plays a factor in it so reduce to lowest equivalent ratio